+++ plugins/ldap/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bs-plugins/ldap
 
+## 1.2.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.16.11
+
 ## 1.2.11
 
 ### Patch Changes